refactor: move old TestBoundingRectangle
